win32 specific code
I haven't been able to make it compile, so don't know if this works, but seems ok for the most part

+++ src/assets.h
@@ -6,7 +6,7 @@ struct Sprite {
     glm::ivec2 atlasOffset;
     glm::ivec2 spriteSize;
 
-    uint atlasIdx;
+    u8 atlasIdx;
 };
 
 enum SpriteID { Player, Chest };
